Refactor: Moved common files into common dir
