+++ src/app.module.ts
@@ -6,9 +6,9 @@ import { AppService } from './app.service';
 import { AuthMiddleware } from './middleware/auth/auth.middleware';
 import { PartyModule } from './party/party.module';
 import { UserModule } from './user/user.module';
-import config from './config/config';
+import config from './common/config';
 import { APP_FILTER } from '@nestjs/core';
-import { HttpExceptionFilter } from './filter/http_exception.filter';
+import { HttpExceptionFilter } from './common/http_exception.filter';
 import { GetUserModule } from './middleware/get_user/get_user.module';
 import { GetUserMiddleware } from './middleware/get_user/get_user.middleware';
 import { ParticipateModule } from './participate/participate.module';
